data_ingestion/requirements.txt
psycopg2-binary
sentence-transformers
PyMuPDF